edit book added

+++ src/app/components/book-list/book-list.component.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 import { BookService } from 'src/app/services/book.service';
+import { DialogsService } from 'src/app/dialogs/dialogs.service';
 
 import { Book } from 'src/app/structures/book';
 
@@ -13,7 +17,10 @@ export class BookListComponent implements OnInit {
 
   books: Book[];
 
-  constructor(private bookService: BookService) {
+  constructor(private bookService: BookService,
+              private router: Router,
+              private dialogsService: DialogsService,
+              private matSnackBar: MatSnackBar) {
   }
 
   ngOnInit() {
@@ -24,13 +31,24 @@ export class BookListComponent implements OnInit {
       });
   }
 
+  editBook(event: Book) {
+    console.log(event.id);
+    this.router.navigate(['book-edit', event.id]);
+  }
+
   deleteBook(event: Book) {
-    this.bookService.delete(event.id)
-      .subscribe(result => {
-        console.log(result);
-        const index = this.books.findIndex(book => book.id === event.id);
-        if (index > -1) {
-          this.books.splice(index, 1);
+    this.dialogsService.confirmDialog('Are you sure?', 'Delete', 'Cancel')
+      .subscribe((confirmed: boolean) => {
+        if (confirmed) {
+          this.bookService.delete(event.id)
+          .subscribe(result => {
+            console.log(result);
+            const index = this.books.findIndex(book => book.id === event.id);
+            if (index > -1) {
+              this.books.splice(index, 1);
+            }
+            this.matSnackBar.open('Book deleted', null, { duration: 3000 });
+          });
         }
       });
   }
